coreSystems/continualTemplating/continualTemplatingGolang/src/test/java/io/continual/templating/impl/golang/GoTemplateEngineTest.java
package io.continual.templating.impl.golang;

import java.io.ByteArrayOutputStream;
import java.io.IOException;
import java.nio.charset.StandardCharsets;

import org.json.JSONObject;

import org.junit.Test;

import io.continual.services.ServiceContainer;
import io.continual.templating.ContinualTemplateContext;
import io.continual.templating.ContinualTemplateEngine.TemplateParseException;
import io.continual.templating.ContinualTemplateSource;
import io.continual.templating.ContinualTemplateSource.TemplateNotFoundException;
import junit.framework.TestCase;

public class GoTemplateEngineTest extends TestCase
{
	@Test
	public void testTemplating () throws IOException, TemplateNotFoundException, TemplateParseException
	{
		final GoTemplateEngine e = new GoTemplateEngine ( new ServiceContainer (), new JSONObject () );

		final ContinualTemplateContext ctc = e.createContext ();
		ctc.put ( "a", 123 );
		ctc.put ( "b", "cde" );

		final ByteArrayOutputStream baos = new ByteArrayOutputStream ();
		e.renderTemplate ( ContinualTemplateSource.fromString ( "{{.a}}b{{.b}}" ), ctc, baos );
		baos.close ();
		final String out = new String ( baos.toByteArray (), StandardCharsets.UTF_8 );
		assertEquals ( "123bcde", out );
	}

	@Test
	public void testTemplatingWithLoops () throws IOException, TemplateNotFoundException, TemplateParseException
	{
		final GoTemplateEngine e = new GoTemplateEngine ( new ServiceContainer (), new JSONObject () );

		final ContinualTemplateContext ctc = e.createContext ();
		ctc.put ( "b", new Foo[] {
			new Foo ( "Fred" ),
			new Foo ( "Jack" ),
			new Foo ( "Bobby" )
		} );
		ctc.put ( "c", "blah" );

		final ByteArrayOutputStream baos = new ByteArrayOutputStream ();
//		e.renderTemplate ( ContinualTemplateSource.fromString ( "{{range .b}}<b>{{.name}}</b>{{end}}" ), ctc, baos );
		e.renderTemplate ( ContinualTemplateSource.fromString ( "{{if .b}}<b>{{.c}}</b>{{end}}" ), ctc, baos );
			// FIXME: the template engine doesn't support "range" :-/
		baos.close ();
		final String out = new String ( baos.toByteArray (), StandardCharsets.UTF_8 );
//		assertEquals ( "<b>Fred</b><b>Jack</b><b>Bobby</b>a", out );
		assertEquals ( "<b>blah</b>", out );
	}

	public static class Foo
	{
		public Foo ( String name ) { fName = name; }
		public String get ( String key ) { return fName; }
		private final String fName;
	}
}
